Move assessment function in the root of the indexables page dir
